requirements.txt

tensorflow
numpy
matplotlib
seaborn
pandas
scikit-learn
librosa
pygame
pillow
ttkbootstrap
